Weapon System Improvements
- Added limited number of uses for weapons
- Added multi shot weapons
- Added AOE effects for damage and slowness

+++ Assets/Scripts/Ant Scripts/AntSO.cs
@@ -10,7 +10,7 @@ public class AntSO : ScriptableObject {
 
 	[Header("Ant Attributes")]
 	public string antName = "";
-	public int moveSpeed = 1;
+	public float moveSpeed = 1.0f;
 	public int jumpHeight = 2;
 	public int health = 100;
